added audio play
we generate audio in rails using elevenlabs text-to-speech.

we fetch the audio in our web ui with 4 params passed on our rails endpoint. rails will use that to get the corresponding content of the talk/thought step.

we play the audio if the audio check box is ticked. otherwise, we do not.

we show the audio check box when NEXT_PUBLIC_ALLOW_AUDIO is true.